pyproject.toml
[project]
name = "heitanbot"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "kamesan1577", email = "kamesan1577@users.noreply.github.com" }
]
dependencies = [
    "discord>=2.3.2",
    "python-dotenv>=1.0.1",
    "quote2image>=0.0.5",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/heitanbot"]
